+++ src/commonMain/kotlin/at/posselt/pfrpg2e/data/events/KingdomEvent.kt
@@ -15,4 +15,5 @@ data class KingdomEvent(
     val traits: Set<KingdomEventTrait>,
     val location: String?,
     val stages: List<KingdomEventStage>,
+    val automationNotes: String?,
 )