+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/dialogs/NewSettlementChoices.kt
@@ -48,7 +48,7 @@ suspend fun newSettlementChoices(terrain: SettlementTerrain? = null): NewSettlem
                 ),
                 Select.fromEnum<SettlementTerrain>(
                     name = "terrain",
-                    label = "Terrain",
+                    label = "Map",
                     value = terrain ?: SettlementTerrain.FOREST,
                 )
             )